completed initial file split and package creation
